feat: begin adding docker image uses patterns
