+++ crates/zizmor/src/audit/known_vulnerable_actions.rs
@@ -13,7 +13,9 @@ use crate::{
     config::Config,
     finding::{Confidence, Finding, Fix, Severity, location::Routable as _},
     github_api,
-    models::{StepCommon, action::CompositeStep, uses::RepositoryUsesExt as _, workflow::Step},
+    models::{
+        StepCommon, action::CompositeStep, uses::repository::RepositoryUsesExt as _, workflow::Step,
+    },
     state::AuditState,
 };
 use yamlpatch::{Op, Patch};
