+++ crates/zizmor/src/models/uses/docker.rs
@@ -1,0 +1,1 @@
+//! Patterns for Docker image `uses:` and corresponding extension traits.
